+++ src/main/java/com/api/healthelp/dao/mapper/PatientMapper.java
@@ -9,14 +9,14 @@ import java.util.List;
 
 public interface PatientMapper {
     List<PatientDTO> getPatients();
-    PatientDTO getPatientById(@Param("patientId")Long id);
-    List<PatientDTO> getPatientsByName(@Param("patientName")String patientName,@Param("userId")Long userId);
-    void insertPatient(@Param("id") Long id, @Param("userId") Long userId,@Param("tenantId") Long tenantId,
+    PatientDTO getPatientById(@Param("patientId")Integer id);
+    List<PatientDTO> getPatientsByName(@Param("patientName")String patientName,@Param("userId")Integer userId);
+    void insertPatient(@Param("id") Integer id, @Param("userId") Integer userId,@Param("tenantId") Integer tenantId,
                           @Param("patientName") String patientName,@Param("patientLastName") String patientLastName,
                           @Param("patientDNI") String patientDNI, @Param("patientTelephone") String patientTelephone,
                           @Param("patientAddress") String patientAddress, @Param("patientLocation") String patientLocation,
                           @Param("patientProfession") String patientProfession, @Param("patientEmail") String patientEmail);
-    void updatePatient(@Param("id") Long id, @Param("userId") Long userId,@Param("tenantId") Long tenantId,
+    void updatePatient(@Param("id") Integer id, @Param("userId") Integer userId,@Param("tenantId") Integer tenantId,
                        @Param("patientName") String patientName,@Param("patientLastName") String patientLastName,
                        @Param("patientDni") String patientDni, @Param("patientTelephone") String patientTelephone,
                        @Param("patientAddress") String patientAddress, @Param("patientLocation") String patientLocation,
